improving remember_me

+++ chapter_10/ex10/remember_me.py
@@ -1,10 +1,13 @@
 from pathlib import Path
 import json
 
-username = input("What is your name? ")
-
 path = Path('username.json')
-contents = json.dumps(username)
-path.write_text(contents)
-
-print(f"We'll remember you when you come back, {username}!")
+if path.exists():
+    contents = path.read_text()
+    username = json.loads(contents)
+    print(f"Welcome back, {username}!")
+else:
+    username = input("What is your name? ")
+    contents = json.dumps(username)
+    path.write_text(contents)
+    print(f"We'll remember you when you come back, {username}!")
